+++ force-app/main/universalLearner/lwc/learnerCourseVisualization/learnerCourseVisualization.html
@@ -1,44 +1,14 @@
 <template>
-    <li class="slds-accordion__list-item">
-        <section class={sectionClass}>
-            <div class="slds-accordion__summary">
-                <h2 class="slds-accordion__summary-heading">
-                    <!--TODO: Put in an aria label for the button-->
-                    <button
-                        class="slds-button slds-button_reset slds-accordion__summary-action"
-                        aria-controls="sectionDetails"
-                        aria-expanded={sectionIsOpen}
-                        title={learnerCourseViewModel.assertion.achievement.name}
-                        onclick={handleSectionHeaderButtonClick}
-                    >
-                        <lightning-icon
-                            icon-name="utility:switch"
-                            size="x-small"
-                            class="slds-accordion__summary-action-icon slds-button__icon slds-button__icon_left"
-                            aria-hidden="true"
-                        ></lightning-icon>
-                        <span class="slds-accordion__summary-content"
-                            >{learnerCourseViewModel.assertion.achievement.name}
-                        </span>
-                    </button>
-                </h2>
-            </div>
-            <div class="slds-accordion__content" hidden={sectionIsHidden} id="sectionDetails">
-                <dl class="slds-list_horizontal slds-wrap">
-                    <template for:each={fieldsToDisplay} for:item="fieldToDisplay">
-                        <dt
-                            key={fieldToDisplay.fieldKey}
-                            title={fieldToDisplay.fieldLabel}
-                            class="slds-item_label slds-text-color_weak slds-truncate"
-                        >
-                            {fieldToDisplay.fieldLabel}:
-                        </dt>
-                        <dd key={fieldToDisplay.fieldKey} class="slds-item_detail slds-truncate">
-                            {fieldToDisplay.fieldValue}
-                        </dd>
-                    </template>
-                </dl>
-            </div>
-        </section>
-    </li>
+    <dl class="slds-list_horizontal slds-wrap">
+        <template for:each={fieldsToDisplay} for:item="fieldToDisplay">
+            <dt
+                key={fieldToDisplay.fieldKey}
+                title={fieldToDisplay.fieldLabel}
+                class="slds-item_label slds-text-color_weak slds-truncate"
+            >
+                {fieldToDisplay.fieldLabel}:
+            </dt>
+            <dd key={fieldToDisplay.fieldKey} class="slds-item_detail slds-truncate">{fieldToDisplay.fieldValue}</dd>
+        </template>
+    </dl>
 </template>
